Before combie HH code

+++ app/src/main/java/org/undp_iwomen/iwomen/ui/fragment/NewIWomenPostFragment.java
@@ -101,11 +101,12 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
 
     //for audio recording process
     private static String mAudioFilePath = null;
-    private MediaPlayer   mPlayer = null;
+    private MediaPlayer mPlayer = null;
     private int timeElapsed = 0, finalTime = 0;
     private Handler durationHandler = new Handler();
     private MediaRecorder mRecorder = null;
     boolean mStartRecording = true;
+    boolean isRecording = false;
     CountDownTimer countDownTimer = null;
 
     //for audio uploading processs
@@ -117,10 +118,16 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
     private SharedPreferences mSharedPreferencesUserInfo;
     private String uploadPhoto;
 
+
+    private final String AUDIO_PERMISSION = "android.permission.RECORD_AUDIO";
+    private final String WRITE_PERMISSION = "android.permission.WRITE_EXTERNAL_STORAGE";
+    private final String READ_PERMISSIOIN = "android.permission.WRITE_EXTERNAL_STORAGE";
+    private final String PREPARE_AUDIO_PERMISSION = "android.permission.MODIFY_AUDIO_SETTINGS";
+
     public NewIWomenPostFragment() {
     }
 
-    public static NewIWomenPostFragment newInstance(){
+    public static NewIWomenPostFragment newInstance() {
         NewIWomenPostFragment fragment = new NewIWomenPostFragment();
         Bundle args = new Bundle();
         fragment.setArguments(args);
@@ -147,13 +154,10 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
         //login_user_id = mSharedPreferencesUserInfo.getString(CommonConfig.USER_ID, null);
 
 
-
-
-
         return rootView;
     }
 
-    public void init(View rootView){
+    public void init(View rootView) {
 
         take_photo_btn = (TextView) rootView.findViewById(R.id.new_post_photo_take_btn);
         upload_photo_btn = (TextView) rootView.findViewById(R.id.new_post_photo_upload_btn);
@@ -194,8 +198,6 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
         }
 
 
-
-
         //setup for progress wheel
         progress_wheel.spin();
         progress_wheel.setRimColor(Color.LTGRAY);
@@ -204,12 +206,12 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
 
     @Override
     public void onClick(View view) {
-        switch (view.getId()){
+        switch (view.getId()) {
             case R.id.new_post_upload_btn:
 
-                if(SKConnectionDetector.getInstance(getActivity()).isConnectingToInternet()){
+                if (SKConnectionDetector.getInstance(getActivity()).isConnectingToInternet()) {
                     checkProcessWhattoDo();
-                }else{
+                } else {
                     SKToastMessage.getInstance(getActivity()).showMessage(getActivity(), "No Internet!", SKToastMessage.ERROR);
                 }
 
@@ -219,14 +221,14 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
             case R.id.new_post_photo_take_btn:
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
                     SKToastMessage.showMessage(getActivity(), getResources().getString(R.string.resource_coming_soon_eng), SKToastMessage.ERROR);
-                }else {
+                } else {
                     takePicture();
                 }
                 break;
 
             case R.id.new_post_photo_upload_btn:
                 //check whether there is permission for READ_STORAGE_PERMISSION
-                if(!hasPermission(CAMERA_PERMISSION)) {
+                if (!hasPermission(CAMERA_PERMISSION)) {
 
                     //if no permission, request permission
                     String[] perms = {CAMERA_PERMISSION};
@@ -235,7 +237,7 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
 
                     requestPermissions(perms, permsRequestCode);
 
-                }else {
+                } else {
 
                     chooseImage();
                 }
@@ -243,8 +245,18 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
 
             case R.id.new_post_audio_upload_btn:
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                    SKToastMessage.showMessage(getActivity(), getResources().getString(R.string.resource_coming_soon_eng), SKToastMessage.ERROR);
-                }else {
+                    if (!hasPermission(AUDIO_PERMISSION) || !hasPermission(WRITE_PERMISSION) || !hasPermission(READ_PERMISSIOIN) || !hasPermission(PREPARE_AUDIO_PERMISSION)) {
+                        String[] perms = {AUDIO_PERMISSION, WRITE_PERMISSION, READ_PERMISSIOIN, PREPARE_AUDIO_PERMISSION};
+
+                        int permsRequestCode = 200;
+
+                        requestPermissions(perms, permsRequestCode);
+                    } else {
+                        performAudioRecord();
+                    }
+
+                    //SKToastMessage.showMessage(getActivity(), getResources().getString(R.string.resource_coming_soon_eng), SKToastMessage.ERROR);
+                } else {
                     performAudioRecord();
                 }
                 break;
@@ -255,23 +267,36 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
                 break;
 
             case R.id.audio_record_done_btn:
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                    SKToastMessage.showMessage(getActivity(), getResources().getString(R.string.resource_coming_soon_eng), SKToastMessage.ERROR);
+
+                if(mAudioFilePath == null || !new File(mAudioFilePath).exists()) {
+                    SKToastMessage.showMessage(getActivity(), "You need to record audio!", SKToastMessage.WARNING);
+                }else if(isRecording){
+                    SKToastMessage.showMessage(getActivity(), "You need to stop recording!", SKToastMessage.WARNING);
                 }else {
+
                     onAudioRecordDoneClick();
+
+                    if (mAudioFilePath == null || !new File(mAudioFilePath).exists()) {
+                        audio_upload_btn.setText(R.string.audio_record_btn_txt);
+                        audio_upload_btn.setBackgroundColor(getResources().getColor(R.color.white));
+                    } else {
+                        audio_upload_btn.setText(R.string.audio_record_done_btn_text);
+                        audio_upload_btn.setBackgroundColor(getResources().getColor(R.color.primary_light));
+                    }
                 }
+
                 break;
 
             case R.id.audio_record_dismiss_btn:
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                    SKToastMessage.showMessage(getActivity(), getResources().getString(R.string.resource_coming_soon_eng), SKToastMessage.ERROR);
-                }else {
-                    onAudioRecordDismissClick();
-                }
+
+                onAudioRecordDismissClick();
+
+                audio_upload_btn.setText(R.string.audio_record_btn_txt);
+                audio_upload_btn.setBackgroundColor(getResources().getColor(R.color.white));
+
                 break;
 
             case R.id.background_audio_record:
-                SKToastMessage.showMessage(getActivity(), getResources().getString(R.string.resource_coming_soon_eng), SKToastMessage.ERROR);
 
                 break;
 
@@ -280,18 +305,24 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
         }
     }
 
-    private void checkProcessWhattoDo(){
+    private void checkProcessWhattoDo() {
 
-        progress_wheel.setVisibility(View.VISIBLE);
 
-        if(crop_file_path != null){
+
+        if (crop_file_path != null) {
             uploadImage();
-        }else if(mAudioFilePath != null){
+        } else if (mAudioFilePath != null) {
             String uploadUrl = "http://api.iwomenapp.org/api/v1/file/audioUpload";
             uploadingAudioFile(uploadUrl, mAudioFilePath);
-        }else{
-            //we will post all data to server, if it's here, we assume all data is ready
-            performPostUpload();
+        } else {
+
+            if(et_postDesc.getText().toString().isEmpty() && crop_file_path == null && mAudioFilePath == null){//this mean user doesn't choose nothing
+                SKToastMessage.showMessage(getActivity(), "To post, you need to add something!", SKToastMessage.WARNING);
+            }else {
+
+                //we will post all data to server, if it's here, we assume all data is ready
+                performPostUpload();
+            }
         }
 
     }
@@ -331,30 +362,32 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
     /***
      * prepare data to upload and post data to server
      */
-    void performPostUpload(){
+    void performPostUpload() {
+
+        progress_wheel.setVisibility(View.VISIBLE);
 
         String content = null, content_type = null, content_mm = null, postUploadName = null;
 
-        if(mstr_lang != null && mstr_lang.equals(Utils.ENG_LANG)){
+        if (mstr_lang != null && mstr_lang.equals(Utils.ENG_LANG)) {
             content = et_postDesc.getText().toString();
-        }else{
+        } else {
             content_mm = et_postDesc.getText().toString();
         }
 
-        if(login_user_name != null) {
+        if (login_user_name != null) {
             postUploadName = login_user_name;
-        }else{
+        } else {
             postUploadName = "";
         }
 
-        if(audio_file_id != null){ //to check if user upload audio
+        if (audio_file_id != null) { //to check if user upload audio
             content_type = "audio";
-        }else{ // user doesn't choose audio nor video
+        } else { // user doesn't choose audio nor video
             content_type = "letter";
         }
 
         NetworkEngine.getInstance().postCreateWomenPost(
-                0, content,
+                1, content,
                 content_type,
                 content_mm,
                 postUploadName,
@@ -388,7 +421,7 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
     };
 
 
-    void performImageCropAfterChose(ChosenImage image){
+    void performImageCropAfterChose(ChosenImage image) {
         progress_wheel.setVisibility(View.GONE);
         if (image != null) {
             //textViewFile.setText(image.getFilePathOriginal());
@@ -405,18 +438,17 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
         }
     }
 
-    public void performAudioRecord(){
+    public void performAudioRecord() {
 
-        if(mAudioFilePath == null || !new File(mAudioFilePath).exists()){
+        if (mAudioFilePath == null || !new File(mAudioFilePath).exists()) {
             mStartRecording = true;
             audio_record_progress_btn.setText("Tap and Talk");
             audio_record_progress_view.setVisibility(View.VISIBLE);
-        }else{
+        } else {
             mStartRecording = false;
             audio_record_progress_btn.setText("Replay");
             audio_record_progress_view.setVisibility(View.VISIBLE);
         }
-
 
 
     }
@@ -453,17 +485,17 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
 
     /***
      * to check where given permission is already permitted or not
+     *
      * @param permission
      * @return
      */
-    private boolean hasPermission(String permission){
-        if(Build.VERSION.SDK_INT>Build.VERSION_CODES.LOLLIPOP_MR1) {
+    private boolean hasPermission(String permission) {
+        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP_MR1) {
             return (getActivity().checkSelfPermission(permission) == PackageManager.PERMISSION_GRANTED);
-        }else{
+        } else {
             return true;
         }
     }
-
 
 
     private void onRecord(boolean start) {
@@ -506,7 +538,7 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
     private Runnable updateProgressTime = new Runnable() {
         public void run() {
 
-            if(mPlayer != null) {
+            if (mPlayer != null) {
                 //get current position
                 timeElapsed = mPlayer.getCurrentPosition();
 
@@ -529,7 +561,7 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
     };
 
     private void stopPlaying() {
-        if(mPlayer != null) {
+        if (mPlayer != null) {
             mPlayer.release();
             mPlayer = null;
         }
@@ -553,7 +585,7 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
     }
 
     private void stopRecording() {
-        if(mRecorder != null) {
+        if (mRecorder != null) {
             mRecorder.stop();
             mRecorder.release();
             mRecorder = null;
@@ -589,38 +621,38 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
                 audio_record_progress_btn.setText("Stop");
 
                 countDownTimer = getCountDown();
-                if(countDownTimer != null){
+                if (countDownTimer != null) {
                     countDownTimer.start();
                 }
 
                 mStartRecording = !mStartRecording;
-
+                isRecording = true;
 
             } else {
 
 
-
-                if(audio_record_progress_btn.getText().equals("Replay")){
+                if (audio_record_progress_btn.getText().equals("Replay")) {
                     audio_record_progress_btn.setText("Stop");
                     onPlay(true);
-                }else{
+                } else {
                     audio_record_progress_btn.setText("Replay");
-                    if(countDownTimer != null){
+                    if (countDownTimer != null) {
                         countDownTimer.cancel();
                     }
                     onPlay(false);
                     stopRecording();
                 }
+                isRecording = false;
             }
         }
     };
 
-    public CountDownTimer getCountDown(){
+    public CountDownTimer getCountDown() {
         return new CountDownTimer(20000, 1000) {
 
             public void onTick(long millisUntilFinished) {
                 //audio_record_progress_btn.setText("Recording : " + millisUntilFinished / 1000);
-                audio_record_progress_bar.setProgress(20 - (int)millisUntilFinished/ 1000);
+                audio_record_progress_bar.setProgress(20 - (int) millisUntilFinished / 1000);
             }
 
             public void onFinish() {
@@ -631,25 +663,28 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
         };
     }
 
-    public void onAudioRecordDoneClick(){
+    public void onAudioRecordDoneClick() {
         audio_record_progress_view.setVisibility(View.INVISIBLE);
         onPlay(false);
         onRecord(false);
 
 
-        if(new File(mAudioFilePath).exists()){
+        if(mAudioFilePath != null)
+        if (new File(mAudioFilePath).exists()) {
             wasAudioRecord = true;
         }
 
     }
 
-    public void onAudioRecordDismissClick(){
+    public void onAudioRecordDismissClick() {
         audio_record_progress_view.setVisibility(View.INVISIBLE);
         onPlay(false);
         onRecord(false);
 
-        File file = new File(mAudioFilePath);
-        boolean deleted = file.delete();
+        if(mAudioFilePath != null) {
+            File file = new File(mAudioFilePath);
+            file.delete();
+        }
     }
 
     private void uploadingAudioFile(String url, String filePath) {
@@ -768,7 +803,7 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
         imageChooserManager.reinitialize(filePath);
     }
 
-    public void uploadImage(){
+    public void uploadImage() {
         File photo = new File(crop_file_path);
         MultipartTypedOutput multipartTypedOutput = new MultipartTypedOutput();
         multipartTypedOutput.addPart("image", new TypedFile("image/png", photo));
@@ -782,7 +817,6 @@ public class NewIWomenPostFragment extends Fragment implements View.OnClickListe
                 }
 
                 checkProcessWhattoDo();
-
             }
 
             @Override
